+++ src/main/java/com/vincevscode/cointracker/config/ApplicationConfiguration.java
@@ -18,6 +18,7 @@ import javax.sql.DataSource;
 import com.vincevscode.cointracker.repository.AuthUserRepositoryInterface;
 import com.vincevscode.cointracker.repository.PostgresAuthUserRepository;
 import com.vincevscode.cointracker.service.AuthUserQueryService;
+import com.vincevscode.cointracker.service.UserManagementService;
 import com.vincevscode.cointracker.service.UserRegistrationService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
@@ -91,5 +92,10 @@ public class ApplicationConfiguration {
             PasswordEncoder passwordEncoder
     ) {
         return new UserRegistrationService(authUserRepository, passwordEncoder);
+    }
+
+    @Bean
+    public UserManagementService userManagementService(AuthUserRepositoryInterface authUserRepository) {
+        return new UserManagementService(authUserRepository);
     }
 }